+++ server/src/api/routes/ShortURL.routes.ts
@@ -3,12 +3,12 @@ import TestApi from "../controllers/Test.controller";
 import CreateShortUrl from "../controllers/CreateShortUrl.controller";
 import RedirectShortUrl from "../controllers/RedirectShortUrl.controller";
 
-const router: Router = Router();
+const ShortUrlRoutes: Router = Router();
 
-router.get("/test", TestApi);
+ShortUrlRoutes.get("/test", TestApi);
 
-router.post("/create", CreateShortUrl);
+ShortUrlRoutes.post("/create", CreateShortUrl);
 
-router.get("/:id", RedirectShortUrl);
+ShortUrlRoutes.get("/:id", RedirectShortUrl);
 
-export default router;
+export default ShortUrlRoutes;